[HW14] Reorganized commands
