--- hw12_13_14_15_calendar/cmd/calendar/cmd/notification/generate.go
@@ -1,15 +1,0 @@
-package notification
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-var generateCmd = &cobra.Command{
-	Use:   "generate",
-	Short: "Generate notifications",
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Generate!")
-	},
-}
