--- hw12_13_14_15_calendar/cmd/calendar/cmd/notification/send.go
@@ -1,14 +1,0 @@
-package notification
-
-import (
-	"fmt"
-	"github.com/spf13/cobra"
-)
-
-var sendCmd = &cobra.Command{
-	Use:   "send",
-	Short: "Send notifications",
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Send!")
-	},
-}
