+++ hw12_13_14_15_calendar/cmd/calendar/cmd/root.go
@@ -1,9 +1,18 @@
 package calendarcmd
 
 import (
-	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/cmd/calendar/cmd/notification"
+	"context"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/cmd/shared"
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/configs"
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/app"
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/logger"
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/server"
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -18,7 +27,47 @@ var (
 		Short: "A simple app to manage your events",
 		Long:  `<Some long desc here...>`,
 		Run: func(cmd *cobra.Command, args []string) {
-			logg.Info("Root command does nothing. Please use special commands")
+			storage := storage.New(cfg.Storage)
+			calendar := app.New(logg, storage)
+
+			server := server.New(logg, calendar, cfg.API)
+
+			ctx, cancel := signal.NotifyContext(context.Background(),
+				syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP)
+			defer cancel()
+
+			go func() {
+				<-ctx.Done()
+
+				ctx, cancel := context.WithTimeout(context.Background(), time.Second*3)
+				defer cancel()
+
+				if err := storage.Dispose(); err != nil {
+					logg.Error("failed to close storage connection: " + err.Error())
+				}
+
+				if err := server.Stop(ctx); err != nil {
+					logg.Error("failed to stop http internal: " + err.Error())
+				}
+			}()
+
+			logg.Info("connecting to storage...")
+
+			if err := storage.Init(ctx); err != nil {
+				logg.Error("failed to connect to storage: " + err.Error())
+				cancel()
+				os.Exit(1)
+			}
+
+			logg.Info("calendar is running...")
+
+			if err := server.Start(ctx); err != nil {
+				logg.Error("failed to start server: " + err.Error())
+				cancel()
+				os.Exit(1)
+			}
+
+			<-ctx.Done()
 		},
 	}
 )
@@ -31,12 +80,10 @@ func Execute() error {
 func init() {
 	cobra.OnInitialize(beforeRun)
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file")
-	rootCmd.AddCommand(notification.Cmd, serverCmd, migrateCmd, versionCmd)
+	rootCmd.AddCommand(migrateCmd, shared.VersionCmd)
 }
 
 func beforeRun() {
 	cfg = configs.New(cfgFile)
 	logg = logger.New(cfg.Logger)
-
-	notification.SetConfig(cfg, logg)
 }
